# secp256k1: Optimize sig verify with mod n scalar.