**This requires #2072**.

This modifies the signature verification function use the new `ModNScalar` type instead of big ints.

This is work towards eventually using the new more efficient mod n scalar throughout.

The following benchmark shows a before and after comparison of typical signature verification:

```
benchmark            old ns/op    new ns/op    delta
------------------------------------------------------
BenchmarkSigVerify   219346       180815       -17.57%

benchmark            old allocs   new allocs   delta
------------------------------------------------------
BenchmarkSigVerify   57           36           -36.84%

benchmark            old bytes    new bytes    delta
------------------------------------------------------
BenchmarkSigVerify   2818         1697         -39.78%
```